# Enhance find_func through trie structure